# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResPlan REST API",
    "description": "Календарно-сетевое и ресурсное планирование проектов: варианты использования UC-1 – UC-9",
    "version": "0.9.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/projects/{id}/tasks/{taskId}": {
      "put": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Изменить задачу: название, длительность, предшественников",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Удалить задачу из сетевого графика",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/projects/{id}/tasks": {
      "get": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Задачи проекта и связи между ними",
        "operationId": "tasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Добавить задачу в сетевой график",
        "operationId": "addTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/conflicts/{id}/resolution": {
      "post": {
        "tags": [
          "UC-6 Разрешить ресурсный конфликт"
        ],
        "summary": "Разрешить конфликт: сохранить одну бронь и снять другую",
        "operationId": "resolve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/conflicts/{id}/escalation": {
      "post": {
        "tags": [
          "UC-6 Разрешить ресурсный конфликт"
        ],
        "summary": "Эскалировать конфликт операционному директору",
        "operationId": "escalate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "tags": [
          "UC-7 Управлять бронированием кандидатов"
        ],
        "summary": "Матрица распределения: действующие брони в периоде",
        "operationId": "list_1",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UC-7 Управлять бронированием кандидатов"
        ],
        "summary": "Забронировать сотрудника на проект (SOFT или HARD)",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResultDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bookings/{id}": {
      "get": {
        "tags": [
          "UC-7 Управлять бронированием кандидатов"
        ],
        "summary": "Бронирование по идентификатору",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "UC-7 Управлять бронированием кандидатов"
        ],
        "summary": "Снять бронь и вернуть сотрудника в пул доступных ресурсов (Bench)",
        "operationId": "release",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "UC-7 Управлять бронированием кандидатов"
        ],
        "summary": "Изменить даты и загрузку или перевести мягкую бронь в жесткую",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResultDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Карточка проекта",
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/schedule": {
      "get": {
        "tags": [
          "UC-1 Управлять расписанием и сетевым графиком"
        ],
        "summary": "Сетевой график: ранние и поздние сроки, резервы, критический путь",
        "operationId": "schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conflicts": {
      "get": {
        "tags": [
          "UC-6 Разрешить ресурсный конфликт"
        ],
        "summary": "Панель ресурсных конфликтов (по умолчанию – OPEN и ESCALATED)",
        "operationId": "list_4",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "OPEN",
                  "RESOLVED",
                  "ESCALATED"
                ]
              },
              "uniqueItems": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConflictDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/conflicts/{id}": {
      "get": {
        "tags": [
          "UC-6 Разрешить ресурсный конфликт"
        ],
        "summary": "Карточка конфликта с пересекающимися бронями и приоритетами проектов",
        "operationId": "get_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TaskRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "durationDays": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 0
          },
          "predecessorIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "required": [
          "durationDays",
          "name"
        ]
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "durationDays": {
            "type": "integer",
            "format": "int32"
          },
          "assigneeId": {
            "type": "integer",
            "format": "int32"
          },
          "assignee": {
            "type": "string"
          },
          "predecessorIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "employee": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "projectPriority": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "loadPercent": {
            "type": "integer",
            "format": "int32"
          },
          "protectedTime": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "BookingResultDto": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConflictDto"
            }
          }
        }
      },
      "ConflictDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "employee": {
            "type": "string"
          },
          "bookingA": {
            "type": "string"
          },
          "bookingB": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "totalLoad": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "first": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "second": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "detectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolutionNote": {
            "type": "string"
          }
        }
      },
...
      "ResolutionDto": {
        "type": "object",
        "properties": {
          "keepBookingId": {
            "type": "integer",
            "format": "int64"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "keepBookingId"
        ]
      },
      "EscalationDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "note"
        ]
      },
      "BookingRequest": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "projectId": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "string",
            "enum": [
              "SOFT",
              "HARD",
              "TRAINING"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "loadPercent": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          }
        },
        "required": [
          "employeeId",
          "endDate",
          "kind",
          "loadPercent",
          "projectId",
          "startDate"
        ]
      },
...
      "BookingPatch": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "SOFT",
              "HARD",
              "TRAINING"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "loadPercent": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          }
        }
      },
...
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "pm": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Row": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "es": {
            "type": "integer",
            "format": "int32"
          },
          "ef": {
            "type": "integer",
            "format": "int32"
          },
          "ls": {
            "type": "integer",
            "format": "int32"
          },
          "lf": {
            "type": "integer",
            "format": "int32"
          },
          "slack": {
            "type": "integer",
            "format": "int32"
          },
          "critical": {
            "type": "boolean"
          },
          "start": {
            "type": "string",
            "format": "date"
          },
          "end": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ScheduleDto": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date"
          },
          "finish": {
            "type": "string",
            "format": "date"
          },
          "durationDays": {
            "type": "integer",
            "format": "int32"
          },
          "criticalPath": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Row"
            }
          }
        }
      },
...
      "ReleaseDto": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/BookingDto"
          },
          "returnedRequestId": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}